+++ src/components/EmptySection.vue
@@ -1,0 +1,14 @@
+<template>
+  <base-warning>
+    <base-warning-title> Ups! Nothing here yet! </base-warning-title>
+    <base-warning-content>
+      <b>This is a great opportunity for you to collaborate!</b> Hit the link at
+      the end of this page and add some examples and a brief description. If you
+      don't know where to start, the
+      <a target="_blank" href="https://docs.python.org/3/library"
+        >Python 3 documentation</a
+      >
+      will lead you in the right direction.
+    </base-warning-content>
+  </base-warning>
+</template>