+++ src/components/layout/TheFooter.vue
@@ -1,20 +1,79 @@
 <script setup lang="ts">
-defineProps<{
+import EditIcon from '~/components/icons/EditIcon.vue'
+import QuestionIcon from '~/components/icons/QuestionIcon.vue'
+import BugIcon from '~/components/icons/BugIcon.vue'
+
+const props = defineProps<{
   repository: string
 }>()
+
+const route = useRoute()
+
+const EditLink = {
+  linkHeader: 'Edit this page on',
+  linkText: 'GitHub',
+  url: `${props.repository}${route.path}.md`,
+  icon: EditIcon,
+}
+
+const footerLinks = [
+  {
+    linkHeader: 'Do you have a question?',
+    linkText: 'ask the community',
+    url: 'https://github.com/wilfredinni/python-cheatsheet/discussions',
+    icon: QuestionIcon,
+  },
+  {
+    linkHeader: 'Do you see a bug?',
+    linkText: 'open an issue on GitHub',
+    url: 'https://github.com/wilfredinni/python-cheatsheet/issues',
+    icon: BugIcon,
+  },
+]
 </script>
 
 <template>
-  <footer
-    class="mt-10 flex items-center justify-between border-t pt-2 dark:border-t-slate-800"
-  >
-    <netlify-badge />
-    <a
-      target="_blank"
-      :href="`${repository}${$route.path}.md`"
-      class="flex items-center space-x-1.5 text-sm text-slate-500 hover:text-sky-500"
-    >
-      <span>Edit this page on github</span>
-    </a>
+  <footer class="mt-5 border-t dark:border-t-slate-800">
+    <div class="justify-between pt-5 sm:flex">
+      <div
+        class="grid text-xs"
+        :class="route.name !== 'index' ? 'space-y-1.5' : ''"
+      >
+        <div
+          v-if="route.name !== 'index'"
+          class="flex items-center text-slate-600 dark:text-slate-400"
+        >
+          <component :is="EditLink.icon" class="mr-2 h-4 w-4" />
+          {{ EditLink.linkHeader }}
+          <a
+            target="_blank"
+            :href="`${repository}${$route.path}.md`"
+            class="ml-1 flex text-sky-500 hover:text-sky-600 dark:text-sky-400 dark:hover:text-sky-500"
+          >
+            {{ EditLink.linkText }}
+          </a>
+        </div>
+
+        <div
+          v-for="link in footerLinks"
+          :key="link.url"
+          class="flex items-center text-slate-600 dark:text-slate-400"
+        >
+          <component :is="link.icon" class="mr-2 h-4 w-4" />
+          {{ link.linkHeader }}
+          <a
+            target="_blank"
+            :href="`${repository}${$route.path}.md`"
+            class="ml-1 flex text-sky-500 hover:text-sky-600 dark:text-sky-400 dark:hover:text-sky-500"
+          >
+            {{ link.linkText }}
+          </a>
+        </div>
+      </div>
+
+      <div class="mt-4 grid sm:mt-0">
+        <netlify-badge class="mx-auto" />
+      </div>
+    </div>
   </footer>
 </template>