+++ src/pages/changelog.md
@@ -7,17 +7,28 @@ updated: July 19, 2022
 
 # Changelog
 
-## [Unreleased](https://github.com/wilfredinni/python-cheatsheet/tree/next)
+<!-- ## [Unreleased](https://github.com/wilfredinni/python-cheatsheet/tree/next) -->
 
-- Add Python **Built-in** Functions.
-- Add Python **standard library** modules:
+## 2022-07-24
+
+### Website
+
+- Added a `contribute` placeholder for unfinished sections.
+- Added useful community links to the footer.
+- Added Pagination.
+
+### Cheatsheet
+
+- Changed titles to be more descriptive.
+- Changed TOC titles to be more concise and descriptive.
+- Added Python **Built-in** Functions.
+- Added Python **standard library** modules:
   - datetime
   - itertools
   - json
   - os
   - pathlib
   - random
-- Add a `contribute` placeholder for unfinished sections.
 
 ## 2022-07-19
 
@@ -28,7 +39,7 @@ updated: July 19, 2022
 The release of the new **pythoncheatsheet.org** website:
 
 - Complete re-design with **Vue 3** and **Tailwind CSS**.
-- The site is **PWA** (Progressive web app) and can be installed in any platform that has a compatible web browser. It is also available offline.
+- The site is now a **PWA** (Progressive web app) and can be installed in any platform that has a compatible web browser.It is also available offline.
 - Added a **dark mode**.
 - Added a **reader mode**.
 - Added **Algolia search** (not available yet).